+++ crates/holochain_data/src/lib.rs
@@ -113,7 +113,7 @@ pub struct HolochainDbConn<I: DatabaseIdentifier> {
 /// # Errors
 ///
 /// Returns an error if `path` is not a directory.
-pub async fn setup_holochain_data<I: DatabaseIdentifier>(
+pub async fn open_db<I: DatabaseIdentifier>(
     path: impl AsRef<Path>,
     database_id: I,
     config: HolochainDataConfig,
@@ -135,9 +135,7 @@ pub async fn setup_holochain_data<I: DatabaseIdentifier>(
 }
 
 #[cfg(feature = "test-utils")]
-pub async fn test_setup_holochain_data<I: DatabaseIdentifier>(
-    database_id: I,
-) -> sqlx::Result<DbWrite<I>> {
+pub async fn test_open_db<I: DatabaseIdentifier>(database_id: I) -> sqlx::Result<DbWrite<I>> {
     let pool = connect_database_memory(HolochainDataConfig::default()).await?;
 
     // Run migrations
@@ -224,7 +222,7 @@ mod tests {
     #[tokio::test]
     async fn in_memory_database_with_migrations() {
         // Set up in-memory database
-        let db = test_setup_holochain_data(TestDbId)
+        let db = test_open_db(TestDbId)
             .await
             .expect("Failed to set up test database");