+++ crates/holochain_state/src/mutations/error.rs
@@ -7,6 +7,9 @@ pub enum StateMutationError {
     Sql(#[from] holochain_sqlite::rusqlite::Error),
 
     #[error(transparent)]
+    Sqlx(#[from] sqlx::Error),
+
+    #[error(transparent)]
     DatabaseError(#[from] holochain_sqlite::error::DatabaseError),
 
     #[error(transparent)]
@@ -29,6 +32,9 @@ pub enum StateMutationError {
 
     #[error("The operation was not found in the cache")]
     OpNotFoundInCache,
+
+    #[error("{0}")]
+    Other(String),
 }
 
 pub type StateMutationResult<T> = Result<T, StateMutationError>;